test: address gemini-code-assist review on PR #4242
- LSMVectorIndexInheritanceTest: drop the try/catch + printStackTrace
  around the CHUNK_EMBEDDING insert (rethrow-only catch is dead code;
  JUnit surfaces the exception directly).
- LSMVectorIndexConcurrentUpdateTest + LSMVectorIndexConcurrentMultiPageUpdateTest:
  split Issue 3135 into its own dedicated class so the integrity-check
  override is scoped to that test only. The pre-existing
  concurrentInsertAndUpdateWithLSMVectorIndex test regains its
  post-test integrity check.
- LSMVectorIndexTest zero-vector regression: replace the manual
  try/catch + boolean flag with AssertJ assertThatThrownBy chained
  isInstanceOf + hasMessageContaining, matching project style.

Verified: engine test-compile clean; 32 tests pass across the four
affected classes.
