+++ engine/src/test/java/com/arcadedb/index/vector/LSMVectorIndexConcurrentUpdateTest.java
@@ -21,7 +21,6 @@ package com.arcadedb.index.vector;
 import com.arcadedb.TestHelper;
 import com.arcadedb.database.DatabaseFactory;
 import com.arcadedb.database.RID;
-import com.arcadedb.exception.DatabaseOperationException;
 import com.arcadedb.index.Index;
 import com.arcadedb.index.TypeIndex;
 import com.arcadedb.query.sql.executor.ResultSet;
@@ -69,19 +68,6 @@ class LSMVectorIndexConcurrentUpdateTest extends TestHelper {
   private static final int CONCURRENT_THREADS = 100;
   private static final int INSERTS_PER_THREAD = 10;
   private static final int UPDATES_PER_THREAD = 20;
-
-  private static final int NUM_RECORDS_3135 = 2000;
-  private static final int CONCURRENT_THREADS_3135 = 100;
-  private static final int EMBEDDING_DIM_3135 = 3072;
-
-  private final AtomicInteger updateErrors3135 = new AtomicInteger(0);
-  private final AtomicInteger verifyErrors3135 = new AtomicInteger(0);
-
-  @Override
-  protected boolean isCheckingDatabaseIntegrity() {
-    // Skip integrity check - we're testing corruption detection
-    return false;
-  }
 
   @Test
   void concurrentInsertAndUpdateWithLSMVectorIndex() throws Exception {
@@ -270,146 +256,6 @@ class LSMVectorIndexConcurrentUpdateTest extends TestHelper {
       for (int i = 0; i < dimensions; i++)
         embedding[i] /= norm;
     }
-    return embedding;
-  }
-
-  // Issue #3135: concurrent UPDATE of multi-page records with LSM_VECTOR index must not corrupt the store.
-  @Test
-  void concurrentMultiPageVectorUpdates() throws Exception {
-    // Phase 1: Create schema with inheritance and LSM_VECTOR index
-    database.transaction(() -> {
-      final Schema schema = database.getSchema();
-
-      // Parent type with embedding (like ContentV in production)
-      final VertexType baseV = schema.createVertexType("BaseV");
-      baseV.createProperty("id", Type.STRING);
-      baseV.createProperty("embedding", Type.ARRAY_OF_FLOATS);
-
-      // Child type (like TextV in production)
-      final VertexType recordV = schema.createVertexType("RecordV");
-      recordV.addSuperType(baseV);  // Inheritance is key!
-
-      // Unique index on id
-      baseV.createTypeIndex(Schema.INDEX_TYPE.LSM_TREE, true, "id");
-
-      // LSM Vector Index on PARENT type
-      database.command("sql", """
-          CREATE INDEX ON BaseV (embedding) LSM_VECTOR
-          METADATA {
-              "dimensions": %d,
-              "similarity": "COSINE"
-          }""".formatted(EMBEDDING_DIM_3135));
-    });
-
-    // Phase 2: Create records with zero embeddings
-    for (int i = 0; i < NUM_RECORDS_3135; i++) {
-      final String id = "record_" + i;
-      final float[] zeroEmbedding = new float[EMBEDDING_DIM_3135];
-      database.transaction(() -> {
-        database.command("sql", "INSERT INTO RecordV SET id=?, embedding=?", id, zeroEmbedding);
-      }, false, 3);
-    }
-
-    // Verify all records were created
-    database.transaction(() -> {
-      try (final ResultSet rs = database.query("sql", "SELECT count(*) as cnt FROM RecordV")) {
-        final long count = rs.hasNext() ? rs.next().<Long>getProperty("cnt") : 0L;
-        assertThat(count).isEqualTo(NUM_RECORDS_3135);
-      }
-    });
-
-    // Phase 3: Concurrently update embeddings
-    final ExecutorService executor = Executors.newFixedThreadPool(CONCURRENT_THREADS_3135);
-    final CountDownLatch latch = new CountDownLatch(NUM_RECORDS_3135);
-
-    for (int i = 0; i < NUM_RECORDS_3135; i++) {
-      final String id = "record_" + i;
-      executor.submit(() -> {
-        try {
-          updateEmbedding3135(id);
-        } catch (final Exception e) {
-          // Log but continue - we're testing for corruption
-        } finally {
-          latch.countDown();
-        }
-      });
-    }
-
-    latch.await(5, TimeUnit.MINUTES);
-    executor.shutdown();
-    executor.awaitTermination(1, TimeUnit.MINUTES);
-
-    // Phase 4: Verification - try to read all embeddings to detect corruption
-    for (int i = 0; i < NUM_RECORDS_3135; i++) {
-      final String id = "record_" + i;
-      try {
-        database.transaction(() -> {
-          try (final ResultSet rs = database.query("sql", "SELECT embedding FROM RecordV WHERE id=?", id)) {
-            if (rs.hasNext()) {
-              final float[] embedding = rs.next().getProperty("embedding");
-              if (embedding == null) {
-                verifyErrors3135.incrementAndGet();
-                System.err.println("   Corruption detected on " + id + ": embedding is null");
-              } else if (embedding.length != EMBEDDING_DIM_3135) {
-                verifyErrors3135.incrementAndGet();
-                System.err.println("   Corruption detected on " + id + ": embedding length is " + embedding.length + " instead of " + EMBEDDING_DIM_3135);
-              }
-            } else {
-              verifyErrors3135.incrementAndGet();
-              System.err.println("   Corruption detected on " + id + ": record not found");
-            }
-          }
-        }, true, 1);
-      } catch (final Exception e) {
-        if (e.getMessage() != null &&
-            (e.getMessage().contains("Invalid pointer") ||
-                e.getMessage().contains("was deleted") ||
-                e.getMessage().contains("Concurrent modification"))) {
-          verifyErrors3135.incrementAndGet();
-          System.err.println("   Corruption detected on " + id + ": " + e.getMessage());
-        } else {
-          throw e;
-        }
-      }
-    }
-
-    // The test should pass with no corruption errors
-    assertThat(verifyErrors3135.get())
-        .as("Corruption errors detected during verification")
-        .isEqualTo(0);
-  }
-
-  private void updateEmbedding3135(final String id) {
-    try {
-      database.transaction(() -> {
-        // Read current record (to simulate real-world read-before-write pattern)
-        try (final ResultSet rs = database.query("sql", "SELECT embedding FROM RecordV WHERE id=?", id)) {
-          if (!rs.hasNext())
-            return;
-          rs.next().getProperty("embedding");
-        }
-
-        // Update with new embedding (multi-page write)
-        final float[] newEmbedding = randomEmbedding3135();
-        database.command("sql", "UPDATE RecordV SET embedding=? WHERE id=?", newEmbedding, id);
-      }, false, 3);
-    } catch (final DatabaseOperationException e) {
-      if (e.getMessage() != null &&
-          (e.getMessage().contains("Invalid pointer") ||
-              e.getMessage().contains("was deleted") ||
-              e.getMessage().contains("Concurrent modification"))) {
-        updateErrors3135.incrementAndGet();
-      } else {
-        throw e;
-      }
-    }
-  }
-
-  private float[] randomEmbedding3135() {
-    final float[] embedding = new float[EMBEDDING_DIM_3135];
-    final ThreadLocalRandom rnd = ThreadLocalRandom.current();
-    for (int i = 0; i < EMBEDDING_DIM_3135; i++)
-      embedding[i] = rnd.nextFloat();
     return embedding;
   }
 }
